Чтение данные передано на Dapper Все таблицы сделаны с полем time типа int64

+++ MetricsAgent/Startup.cs
@@ -65,7 +65,7 @@ namespace MetricsAgent
                 command.ExecuteNonQuery();
 
                 command.CommandText = @"CREATE TABLE dotnetmetrics(id INTEGER PRIMARY KEY,
-                    value INT, time INT)"; ;
+                    value INT, time INT64)"; ;
                 // отправляем запрос в базу данных
                 command.ExecuteNonQuery();
 
@@ -75,7 +75,7 @@ namespace MetricsAgent
                 command.ExecuteNonQuery();
 
                 command.CommandText = @"CREATE TABLE hddmetrics(id INTEGER PRIMARY KEY,
-                    value INT, time INT)"; ;
+                    value INT, time INT64)"; ;
                 // отправляем запрос в базу данных
                 command.ExecuteNonQuery();
 
@@ -85,7 +85,7 @@ namespace MetricsAgent
                 command.ExecuteNonQuery();
 
                 command.CommandText = @"CREATE TABLE networkmetrics(id INTEGER PRIMARY KEY,
-                    value INT, time INT)"; ;
+                    value INT, time INT64)"; ;
                 // отправляем запрос в базу данных
                 command.ExecuteNonQuery();
 
@@ -95,7 +95,7 @@ namespace MetricsAgent
                 command.ExecuteNonQuery();
 
                 command.CommandText = @"CREATE TABLE rammetrics(id INTEGER PRIMARY KEY,
-                    value INT, time INT)"; ;
+                    value INT, time INT64)"; ;
                 // отправляем запрос в базу данных
                 command.ExecuteNonQuery();
             }
